add q3k's FMISS VM (read id works!) q3k implemented some of the S5L8720's FMISS VM and I massaged it into this. as far as I can tell, read id works correctly. it even attempts to start reading data from the device but, obviously, that's not going to work (yet).
this is a huge step in the right direction for the 16gb NAND project because I'd like some visibility into what the iPod's doing inside the NAND driver more than four bytes through nand_lba at a time!

I also added ipod/ipod-debug which is literally just a program that exists to define debug symbols for GDB. in the init file there's a dictionary of where known static datastructures are in memory (or, at least, static pointers to them). the FIL function table looks incorrect which is fascinating since it looks right in Ghidra, but that's for another time. doesn't seem super important right now.