NAND Initialization Works!
AND_Init - and I assume just reading from NAND in general - works. The emulator makes its way to trying to load the Disk Mode binary and then the CPU has an exception for some reason. I'm not sure why it's immediately going to Disk Mode. AES isn't being invoked and I assume that's what needs to happen for RetailOS to load (as I assume it's encrypted). I'm not sure. I think the next thing to do is to get the clickwheel working and see if diag works.

I also removed the ADM since that's likely equivalent to the FMISS layer that's included in the NAND peripheral for better or for worse.